object_lambda_benchmark/use_cases/ptr/object_lambda/word_counter/requirements.txt
requests
aioboto3
pandas
